fix: refactor some
